==> src/2_extract_and_concat.ipynb <==
## AIS数据按每小时取平均，并且将所有数据合并到一个csv文件

import pandas as pd
import numpy as np
import movingpandas as mdp
from datetime import datetime


# 新的数据容器
df_processed = pd.DataFrame(data=None,columns=['SOG','heading','draught', 'wind_val', 'wind_direction', 'wave_val', 'wave_direction', 'stream_val', 'stream_direction', 'postime', 'num'])

for shipnum in range(1, 14): # 13条船

    # 读取一条船跑满24小时的AIS数据
    df = pd.read_csv('processed_%d.csv' % shipnum,index_col=0)
    df['postime'] = pd.to_datetime(df['postime'],format='ISO8601')

    # 读取noon report，并且筛选出年月日
    df_fuel = pd.read_csv('../raw_data/mrv.csv')
    # 时间转换成datetime对象
    df_fuel['input_date'] = pd.to_datetime(df_fuel['input_date'],format='ISO8601')
    # 选择船的编号
    df_fuel = df_fuel[df_fuel['num'] == shipnum]
    # 筛选出一直在跑的船，这边考虑daily_voyage_time_mrv大于等于24小时的轨迹
    df_fuel = df_fuel[df_fuel['daily_voyage_time_mrv'] >= 24]
    # 删除关键字段存在空值的行，daily_fuel_mrv
    df_fuel = df_fuel.dropna(axis=0, subset = ['daily_fuel_mrv'])
    # 序号重新排列
    df_fuel = df_fuel.reset_index(drop=True)
    # 将日期全部抽取出来
    years = df_fuel['input_date'].dt.year
    months = df_fuel['input_date'].dt.month
    days = df_fuel['input_date'].dt.day

    # 临时数据容器，储存一条船3年的数据（经过每小时平均的）
    temp_data = pd.DataFrame(data=None,columns=['SOG','heading','draught', 'wind_val', 'wind_direction', 'wave_val', 'wave_direction', 'stream_val', 'stream_direction', 'postime', 'num'])
    for i in range(df_fuel.shape[0]): # 遍历一条船所有跑满24小时的天数
        for hours in range(24): #遍历每个小时内的点，并且做平均
            # 同一个小时的所有点
            temp_hours = df[(df['postime'].dt.year ==years[i]) & (df['postime'].dt.month ==months[i]) & (df['postime'].dt.day ==days[i]) & (df['postime'].dt.hour == hours)]
            avg = temp_hours[['SOG','heading','draught',  'wind_val', 'wind_direction', 'wave_val', 'wave_direction', 'stream_val', 'stream_direction']].mean()
            avg['postime'] = datetime(years[i], months[i], days[i], hours, 0, 0)
            avg['num'] = shipnum
            avg = avg.to_frame(); avg = avg.transpose()
            temp_data = pd.concat([temp_data, avg], ignore_index = True)
        temp_data = temp_data.dropna(axis=0)
    df_processed = pd.concat([df_processed, temp_data], ignore_index = True)

df_processed

df_processed.to_csv('AIS_hour.csv')
