==> src/5_training.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import os

from sklearn.model_selection import train_test_split
from sklearn.metrics import mean_squared_error
from sklearn.preprocessing import StandardScaler

from xgboost import XGBRegressor
from sklearn.model_selection import GridSearchCV
from sklearn.model_selection import ShuffleSplit
import autoxgb
from xgboost import plot_importance


df = pd.read_csv('../Training_data/data_13_17_1_std.csv')
df = df.drop(['Unnamed: 0','Pe', 'R_total', 'R_calm', 'postime', 'num'], axis = 1)

X = df.drop(['R_added'], axis = 1)
Y = df[['R_added']]

from sklearn.preprocessing import StandardScaler
# 实例化
ss = StandardScaler()
# 特征数据
X = ss.fit_transform(X)
# 目标变量
y = Y
# 先切分数据集
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=9)

from sklearn.tree import DecisionTreeRegressor

tr = DecisionTreeRegressor(max_depth=10)

tr.fit(X_train, y_train)
# 预测值
tr_pre = tr.predict(X_test)
tr_pre =tr_pre.reshape(-1,1)
# 模型评分
print('Score:{:.4f}'.format(tr.score(X_test, y_test)))
# RMSE(标准误差)
print('RMSE:{:.4f}'.format(np.sqrt(mean_squared_error(y_test,tr_pre))))

test_pre = pd.DataFrame({"test" : y_test['R_added'].tolist(),"pre": tr_pre.flatten()})
test_pre

from sklearn.svm import SVR

linear_svr = SVR(kernel="linear")
linear_svr.fit(X_train, y_train)
linear_svr_pre = linear_svr.predict(X_test)

# 模型评分
print('Score:{:.4f}'.format(linear_svr.score(X_test, y_test)))
# RMSE(标准误差)
print('RMSE:{:.4f}'.format(np.sqrt(mean_squared_error(y_test,linear_svr_pre))))

from sklearn import ensemble

gb = ensemble.GradientBoostingRegressor()

gb.fit(X_train, y_train)
gb_pre=gb.predict(X_test)

# 模型评分
print('Score:{:.4f}'.format(gb.score(X_test, y_test)))
# RMSE(标准误差)
print('RMSE:{:.4f}'.format(np.sqrt(mean_squared_error(y_test,gb_pre))))

from xgboost import XGBRegressor
from sklearn.model_selection import GridSearchCV
from sklearn.model_selection import ShuffleSplit

from xgboost import plot_importance

xgb_model = XGBRegressor(nthread=7)
cv_split = ShuffleSplit(n_splits=6, train_size=0.7, test_size=0.2)
grid_params = dict(
    max_depth = [4, 5, 6, 7],
    learning_rate = np.linspace(0.03, 0.3, 10),
    n_estimators = [100, 200]
)
grid_model = GridSearchCV(xgb_model, grid_params, cv=cv_split, scoring='neg_mean_squared_error')
grid_model.fit(X_train, y_train)

pred = grid_model.predict(X_test)

print('RMSE:{:.4f}'.format(np.sqrt(mean_squared_error(y_test,pred))))

test_pre = pd.DataFrame({"test" : y_test['R_added'].tolist(),"pre": pred.flatten()})

test_pre

### Use AutoXGB

# X_train, X_test, y_train, y_test

df = pd.read_csv('../Training_data/data_13_17_1_std.csv')
df = df.drop(['Unnamed: 0','Pe', 'R_total', 'R_calm', 'postime', 'num'], axis = 1)

msk = np.random.rand(len(df)) < 0.8

train = df[msk]

test = df[~msk]

train.to_csv('train_auxgb.csv')
test.to_csv('test_auxgb.csv')



from autoxgb import AutoXGB


# required parameters:
train_filename = "train_auxgb.csv"
output = "output"

# optional parameters
test_filename = 'test_auxgb.csv'
task = 'regression'
idx = None
targets = ["R_added"]
features = None
# categorical_features = ['SOG', 'heading', 'draught', 'wind_val', 'wind_direction', ]
categorical_features = train.columns
use_gpu = False
num_folds = 5
seed = 42
num_trials = 100
time_limit = 360
fast = fast

# Now its time to train the model!
axgb = AutoXGB(
    train_filename=train_filename,
    output=output,
    test_filename=test_filename,
    task=task,
    idx=idx,
    targets=targets,
    features=features,
    categorical_features=categorical_features,
    use_gpu=use_gpu,
    num_folds=num_folds,
    seed=seed,
    num_trials=num_trials,
    time_limit=time_limit,
    fast=fast,
)
axgb.train()
